requirements.txt
# https://pip.pypa.io/en/stable/reference/pip_install/#git
# https://pip.pypa.io/en/latest/reference/pip_install/#requirements-file-format
# https://packaging.python.org/discussions/install-requires-vs-requirements/

# atlassian-python-api==1.15.9
git+https://github.com/atlassian-api/atlassian-python-api.git@5b860b444283cd3fb7eddc5659996db9a5165a89

Pygments==2.6.1
mistune==2.0.0a4
urllib3==1.25.9
pyyaml==5.3.1

pickleDB==0.9.2
